+++ src/frontend/apps/impress/src/api/index.ts
@@ -1,6 +1,7 @@
 export * from './APIError';
 export * from './config';
 export * from './fetchApi';
+export * from './gristApi';
 export * from './helpers';
 export * from './types';
 export * from './utils';